move `GetPerfectPerformance` to `PerformanceCalculator`
